+++ src/Model/Asn.php
@@ -20,24 +20,33 @@ use GeoIp2\Util;
  *      the record. In particular, this is the largest network where all of the
  *      fields besides $ipAddress have the same value.
  */
-class Asn extends AbstractModel
+class Asn implements \JsonSerializable
 {
-    protected ?int $autonomousSystemNumber;
-    protected ?string $autonomousSystemOrganization;
-    protected string $ipAddress;
-    protected string $network;
+    public readonly ?int $autonomousSystemNumber;
+    public readonly ?string $autonomousSystemOrganization;
+    public readonly string $ipAddress;
+    public readonly string $network;
 
     /**
      * @ignore
      */
     public function __construct(array $raw)
     {
-        parent::__construct($raw);
-        $this->autonomousSystemNumber = $this->get('autonomous_system_number');
+        $this->autonomousSystemNumber = $raw['autonomous_system_number'] ?? null;
         $this->autonomousSystemOrganization =
-            $this->get('autonomous_system_organization');
-        $ipAddress = $this->get('ip_address');
+            $raw['autonomous_system_organization'] ?? null;
+        $ipAddress = $raw['ip_address'];
         $this->ipAddress = $ipAddress;
-        $this->network = Util::cidr($ipAddress, $this->get('prefix_len'));
+        $this->network = Util::cidr($ipAddress, $raw['prefix_len']);
+    }
+
+    public function jsonSerialize(): ?array
+    {
+        return [
+            'autonomous_system_number' => $this->autonomousSystemNumber,
+            'autonomous_system_organization' => $this->autonomousSystemOrganization,
+            'ip_address' => $this->ipAddress,
+            'network' => $this->network,
+        ];
     }
 }