+++ tests/bootstrap.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-if (!$loader = @include __DIR__ . '/../vendor/autoload.php') {
-    exit('Project dependencies missing');
-}
-
-$loader->add('GeoIp2\Test', __DIR__);