+++ CHANGELOG.md
@@ -5,6 +5,11 @@ CHANGELOG
 -------------------
 
 * IMPORTANT: PHP 8.1 or greater is now required.
+* BREAKING: Read-only properties are now used for the model and record
+  classes rather than magic methods. This significantly improves performance.
+* BREAKING: The `raw` property on model classess and the `record` property on
+  record classes have been removed.
+* BREAKING: The `jsonSerialize` output has changed.
 * `GeoIp2\WebService\Client` methods now throw an `InvalidArgumentException`
   if an invalid IP address is passed to them. Previously, they would make
   a request to the web service and throw a